Added CMS functionalities to the Astro app
The Astro app is updated with several changes to its CMS functionalities. It includes the addition of new models for the hero section on the main page, featured categories and featured items. These models cover how these items are stored and updated. Methods for retrieving these items were added to the SyncService. Quick search functionality was added to the product controller, and a search filter was applied to only return active items. Overall, these changes provide a backbone for organizing key website features and for quick and efficient site-wide searches.